refactor: scope support for recovery

+++ routes/remove_recovery_record.go
@@ -4,11 +4,13 @@ import (
 	"fmt"
 	"os"
 	"strconv"
+	"strings"
 
 	"github.com/MertJSX/folder-host-go/database/logs"
 	"github.com/MertJSX/folder-host-go/database/recovery"
 	"github.com/MertJSX/folder-host-go/types"
 	"github.com/MertJSX/folder-host-go/utils"
+	"github.com/MertJSX/folder-host-go/utils/config"
 	"github.com/gofiber/fiber/v2"
 )
 
@@ -27,6 +29,7 @@ func RemoveRecoveryRecord(c *fiber.Ctx) error {
 		})
 	}
 
+	account := c.Locals("account").(types.Account)
 	currentRecord, err := recovery.GetRecoveryRecord(idToInt)
 
 	if err != nil {
@@ -41,6 +44,12 @@ func RemoveRecoveryRecord(c *fiber.Ctx) error {
 		isExistingItem = false
 	}
 
+	if !strings.HasPrefix(currentRecord.OldLocation, config.Config.Folder+account.Scope) {
+		return c.Status(403).JSON(
+			fiber.Map{"err": "Out of scope error! No permission!"},
+		)
+	}
+
 	if isExistingItem {
 		if err = os.RemoveAll(currentRecord.BinLocation); err != nil {
 			return c.Status(500).JSON(fiber.Map{
@@ -49,7 +58,7 @@ func RemoveRecoveryRecord(c *fiber.Ctx) error {
 		}
 	}
 
-	err = recovery.DeleteRecoveryRecord(idToInt)
+	err = recovery.DeleteRecoveryRecord(idToInt, config.Config.Folder+account.Scope)
 
 	if err != nil {
 		return c.Status(500).JSON(fiber.Map{
